requirement.txt
asgiref==3.8.1
boto3==1.35.13
botocore==1.35.13
Django==5.1.1
django-autoslug==1.9.9
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
progressbar2==4.5.0
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
